Update ansible-test support for CI providers.
Refactored CI provider code to simplify multiple provider support and addition of new providers.
