+++ changelogs/fragments/ansible-test-ci-support.yml
@@ -1,0 +1,2 @@
+minor_changes:
+    - ansible-test - Refactored CI related logic into a basic provider abstraction.
