+++ test/lib/ansible_test/_internal/cloud/azure.py
@@ -11,7 +11,6 @@ from ..io import (
 from ..util import (
     ApplicationError,
     display,
-    is_shippable,
     ConfigParser,
 )
 
@@ -55,13 +54,10 @@ class AzureCloudProvider(CloudProvider):
 
         aci = self._create_ansible_core_ci()
 
-        if os.path.isfile(aci.ci_key):
+        if aci.available:
             return
 
         if os.path.isfile(self.SHERLOCK_CONFIG_PATH):
-            return
-
-        if is_shippable():
             return
 
         super(AzureCloudProvider, self).filter(targets, exclude)
