+++ test/lib/ansible_test/_internal/cloud/tower.py
@@ -8,7 +8,6 @@ import time
 from ..util import (
     display,
     ApplicationError,
-    is_shippable,
     SubprocessError,
     ConfigParser,
 )
@@ -49,10 +48,7 @@ class TowerCloudProvider(CloudProvider):
 
         aci = get_tower_aci(self.args)
 
-        if os.path.isfile(aci.ci_key):
-            return
-
-        if is_shippable():
+        if aci.available:
             return
 
         super(TowerCloudProvider, self).filter(targets, exclude)
