+++ test/lib/ansible_test/_internal/config.py
@@ -8,7 +8,6 @@ import sys
 from . import types as t
 
 from .util import (
-    is_shippable,
     docker_qualify_image,
     find_python,
     generate_pip_command,
@@ -242,17 +241,7 @@ class SanityConfig(TestConfig):
         self.list_tests = args.list_tests  # type: bool
         self.allow_disabled = args.allow_disabled  # type: bool
         self.enable_optional_errors = args.enable_optional_errors  # type: bool
-
-        if args.base_branch:
-            self.base_branch = args.base_branch  # str
-        elif is_shippable():
-            self.base_branch = os.environ.get('BASE_BRANCH', '')  # str
-
-            if self.base_branch:
-                self.base_branch = 'origin/%s' % self.base_branch
-        else:
-            self.base_branch = ''
-
+        self.base_branch = args.base_branch  # type: str
         self.info_stderr = self.lint
