+++ test/lib/ansible_test/_internal/core_ci.py
@@ -28,7 +28,6 @@ from .io import (
 from .util import (
     ApplicationError,
     display,
-    is_shippable,
     ANSIBLE_TEST_DATA_ROOT,
 )
 
@@ -39,6 +38,11 @@ from .util_common import (
 
 from .config import (
     EnvironmentConfig,
+)
+
+from .ci import (
+    AuthContext,
+    get_ci_provider,
 )
 
 from .data import (
@@ -75,13 +79,14 @@ class AnsibleCoreCI:
         self.endpoint = None
         self.max_threshold = 1
         self.retries = 3
+        self.ci_provider = get_ci_provider()
+        self.auth_context = AuthContext()
 
         if self.arch:
             self.name = '%s-%s-%s' % (self.arch, self.platform, self.version)
         else:
             self.name = '%s-%s' % (self.platform, self.version)
 
-        self.ci_key = os.path.expanduser('~/.ansible-core-ci.key')
         self.resource = 'jobs'
 
         # Assign each supported platform to one provider.
@@ -159,7 +164,7 @@ class AnsibleCoreCI:
                 # permit command-line override of region selection
                 region = args.remote_aws_region
                 # use a dedicated CI key when overriding the region selection
-                self.ci_key += '.%s' % args.remote_aws_region
+                self.auth_context.region = args.remote_aws_region
             else:
                 region = 'us-east-1'
 
@@ -243,6 +248,11 @@ class AnsibleCoreCI:
 
         raise ApplicationError('Unable to get available endpoints.')
 
+    @property
+    def available(self):
+        """Return True if Ansible Core CI is supported."""
+        return self.ci_provider.supports_core_ci_auth(self.auth_context)
+
     def start(self):
         """Start instance."""
         if self.started:
@@ -250,30 +260,7 @@ class AnsibleCoreCI:
                          verbosity=1)
             return None
 
-        if is_shippable():
-            return self.start_shippable()
-
-        return self.start_remote()
-
-    def start_remote(self):
-        """Start instance for remote development/testing."""
-        auth_key = read_text_file(self.ci_key).strip()
-
-        return self._start(dict(
-            remote=dict(
-                key=auth_key,
-                nonce=None,
-            ),
-        ))
-
-    def start_shippable(self):
-        """Start instance on Shippable."""
-        return self._start(dict(
-            shippable=dict(
-                run_id=os.environ['SHIPPABLE_BUILD_ID'],
-                job_number=int(os.environ['SHIPPABLE_JOB_NUMBER']),
-            ),
-        ))
+        return self._start(self.ci_provider.prepare_core_ci_auth(self.auth_context))
 
     def stop(self):
         """Stop instance."""
